Fix clipped log format without newline
Without this patch, this log file has no newline, so the log is one
very long line. (And the contentType and userAgent are missing.)

(I'm unclear about whether this originated with the template or if it
was from the file distributed with rundeck itself; at any rate, it seems
to be correct upstream).