Enable "roleUsernameMemberAttribute" for LDAP auth
For sites using RFC 2307 posixGroup for roles, the *roleMemberAttribute*
does not work, because the posixGroup *memberUid* attributes are
usernames and not DNs; the *roleUsernameMemberAttribute* is for this.

Since *role_member_attribute* is already specified in params, it is
retained as the default unless *role_username_member_attribute* is set.
I think these two are mutually exclusive, so there should be no case
where they would both be enabled.