Add support for setting jvm_args
With $jvm_args only accessible in rundeck::params, it is not possible
override the values there. It is necessary add to the JVM parameters to
configure rundeckd to listen on a different port than the default and
bind to a specific IP address, both of which are preferred when setting
up Apache httpd to proxy (I want to change the port to one that SELinux
allows httpd to connect to by default, such as 8008); it is also
necessary to add "-Drundeck.jetty.connector.forwarded=true" to enable it
to use the X-Forwarded-* HTTP headers.

I am also just replacing the default from rundeck::params, rather than
merging or concatenating, as I can easily imagine wanting to use
different memory parameters than the ones hard-coded there.