Fix for #100 deep merging issue with aclpolicyfile
Updated the parameter defaults to fixed values not inherited from the
root rundeck class because it avoids any issues if the define is used on
it's own (not sure why you would do that) and also allows me to write a
test for the default case of the properties.dir setting.

Due to the above change I also update the config.pp file accordingly to
pass-through the remaining parameters.