A new function to validate the acl policies
Given the complexities of writing rundeck's acl policies it has come up
again and again where passing incorrect data into the hash can cause
puppet to generate an invalid policy file.

The point of this change is to restructure the hash so that it better
repsents the structure of the acl policy itself and to provide a parser
function for the policy hash so that the data can be validated.