Use FQDN based random value for 'rundeck.server.uuid' when 'serialnumber' fact is not present
When the 'serialnumber' fact is not present the 'rundeck.server.uuid'
setting is set to 'undef'. Also when the 'strict_variables' setting
is set to true, the Puppet parser will fail if the 'serialnumber' fact
is not present.

When the 'serialnumber' fact is present it will still use it, if not it
will fallback to using a random value based on the FQDN. Therefor this
change should not break existing installations.