Don't manage user / group by default
Add new parameters (manage_user and manage_group) to control whether the
Rundeck user is managed.
Now that user_id and group_id are optional integers, we can simplify this code
somewhat. To make the ordering work properly, and for consistency, this also
moves the file resources to config manifest.